重构: 将 fileheader_utf 结构体和适配的转换函数抽取为工具公共代码
